+++ src/shared/components/palette/case-flag/domain/case-flag-state.model.ts
@@ -12,4 +12,5 @@ export interface CaseFlagState {
   flagCode?: string;
   listOfValues?: { key: string, value: string }[];
   selectedFlagDetail?: FlagDetail;
+  flagsCaseFieldId?: string;
 }